# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Category API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/qr/create": {
            "post": {
                "tags": [
                    "QR Code"
                ],
                "summary": "Create a new QR code",
                "operationId": "13f999bf61f2169340f9ef97d1b953bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "restaurantId": {
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "tableNo": {
                                        "type": "integer",
                                        "example": "12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QR code generated and stored successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/qr/{id}": {
            "get": {
                "tags": [
                    "QR Code"
                ],
                "summary": "Get a QR code by ID",
                "operationId": "5255116da45d09b44d723a372e22069a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code data"
                    },
                    "404": {
                        "description": "QR code not found"
                    }
                }
            }
        },
        "/qr/update/{id}": {
            "put": {
                "tags": [
                    "QR Code"
                ],
                "summary": "Update a QR code by ID",
                "operationId": "5291d9f3ce3cd300bcf6adcc00a17584",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "restaurantId": {
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "tableNo": {
                                        "type": "integer",
                                        "example": "12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QR code updated successfully"
                    },
                    "404": {
                        "description": "QR code not found"
                    }
                }
            }
        },
        "/qr/delete/{id}": {
            "delete": {
                "tags": [
                    "QR Code"
                ],
                "summary": "Delete a QR code by ID",
                "operationId": "81f07a28c3c07911d60ace77816b13cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code deleted successfully"
                    },
                    "404": {
                        "description": "QR code not found"
                    }
                }
            }
        },
        "/reports/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Dashboard Statistics for a Specific Restaurant",
                "description": "Fetches the dashboard statistics for a specific restaurant, including today's collection, total invoices, completed orders, and rejected orders.",
                "operationId": "e12d063b9dc00384af305459ed38f4d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Restaurant ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "todayCollection": {
                                            "type": "number",
                                            "example": 1234.56
                                        },
                                        "totalInvoiceToday": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "totalCompleteOrder": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "totalRejectOrder": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Restaurant Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}